add car lookup by proximity to a [lng,lat]
The car params schema now supports lat,lng and radius (in km) to filter cars based on their proximity
to a sepecifc point on earth. validation rules are applied to these new fields.

There is a new util module called map that will export all the map related functionality.
We are using a geoencoding api called opencage,
the api key is added to the .env file.
If you are a developer, please contact me to get the api key for yourself, which you have to store
securely on your machine.

Car model was updated to add the coupe type to cars.
the sample data was updated to accomodate location lookups.

Signed-off-by: Hassan Attar <info@hassanattar.com>